Fix integer types

+++ src/common/DataBlockAllocator.h
@@ -24,13 +24,13 @@ class CGUID;
 struct tagDBOpParam
 {
 	//当前读写数据块编号
-	ulong	nCurNum;
+	uint32_t	nCurNum;
 	//当前读写数据块的大小
-	ulong	nCurDBSize;
+	uint32_t	nCurDBSize;
 	//当前读写数据的位置
-	ulong	nCurPos;
+	uint32_t	nCurPos;
 	//当前读写数据块的指针
-	uchar	*pDBPtr;
+	byte	*pDBPtr;
 };
 
 class CDataBlockAllocator
@@ -50,13 +50,13 @@ private:
 	const int	m_nDBSize;
 	//用在逻辑层临时使用的数据块
 	//个数不多，使用较大的数据块来满足各种大小的需求
-    typedef std::multimap<long, uchar*>	MMapShareDB;
-    typedef std::pair <long, uchar*> ShareDBPair;
-    typedef std::multimap<long, uchar*>::iterator itShareDB;
+    typedef std::multimap<long, byte*>	MMapShareDB;
+    typedef std::pair <long, byte*> ShareDBPair;
+    typedef std::multimap<long, byte*>::iterator itShareDB;
 	MMapShareDB m_ShareDataBlocks;
 
-    typedef std::map<uchar*, long>	mapAllocedRecord;
-    typedef std::map<uchar*, long>::iterator itAllocR;
+    typedef std::map<byte*, long>	mapAllocedRecord;
+    typedef std::map<byte*, long>::iterator itAllocR;
 	mapAllocedRecord	m_AllocedShareDB;
 	//允许的最大共享数据块个数
     std::map<long, long>	m_MapTest;
@@ -68,8 +68,8 @@ public:
 	CDataBlock*	AllocDB(long lTestFlag=0);
 	void FreeDB(CDataBlock*);
 
-	uchar* AllockShareDB(long lSize);
-	void FreeShareDB(uchar* pData);
+	byte* AllockShareDB(long lSize);
+	void FreeShareDB(byte* pData);
 
 	void PutAllocInfo(const char* pszFileName);
 };
